backend/django/requirements.txt
beautifulsoup4==4.12.3
celery==5.3.6
Django==4.2.11
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.2
django-redis==5.4.0
django-sql-explorer==4.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==21.2.0
pika==1.3.2
pillow==10.2.0
psycopg2-binary==2.9.9
redis==5.0.3